+++ src/main/java/com/likelion12th/SWUProject1Team/repository/ResumeRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ResumeRepository extends JpaRepository<Resume, Long> {
+public interface ResumeRepository extends JpaRepository<Resume, Integer> {
     List<Resume> findByMember_MemberId(Integer memberId);
 }
