+++ src/main/java/com/likelion12th/SWUProject1Team/repository/WorkExperienceRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface WorkExperienceRepository extends JpaRepository<WorkExperience, Long> {
-    List<WorkExperience> findByResumeId(Long resumeId); // resumeId로 WorkExperience 찾기
+public interface WorkExperienceRepository extends JpaRepository<WorkExperience, Integer> {
+    List<WorkExperience> findByResumeId(Integer resumeId); // resumeId로 WorkExperience 찾기
 
-    List<WorkExperience> findByResume_Id(Long resumeId);
+    List<WorkExperience> findByResume_Id(Integer resumeId);
 }
